\name{key}
\alias{key}
\docType{data}
\title{
Key of Items 
}
\description{
This data set contains the correct responses of the items in the test data set.
}
\usage{data("key")}
\format{
  A data frame with variables \code{Question01 - Question40}.
}
\author{Michela Battauz}
\seealso{
	\code{\link{addkey}}
}
\examples{
data(key)
key
}
\keyword{datasets}
