\name{weights}
\alias{weights}
\docType{data}
\title{
Weights of Items 
}
\description{
This data set contains the weights of the correct responses of the items in the test data set.
}
\usage{data("weights")}
\format{
  A data frame with variables \code{Question01 - Question40}.
}
\author{Michela Battauz}
\seealso{
	\code{\link{addweights}}
}
\examples{
data(weights)
weights
}
\keyword{datasets}
