\name{key}
\alias{key}
\docType{data}
\title{
Key of Items 
}
\description{
This data set contains the correct responses of the items of all data sets of the fsia package.
}
\usage{data("key")}
\format{
  A data frame with variables \code{I1-I50}.
}
\author{Michela Battauz}
\seealso{
	\code{\link{exam1}}, \code{\link{mchoice1}}, \code{\link{mchoice1id}}, \code{\link{mchoice2}}, \code{\link{mchoice2id}}
}
\examples{
data(key)
key
}
\keyword{datasets}
